Alloy version bump and removal of `alloy-signer-local` (#3868)
## Motivation

The versions of alloy used in examples/Cargo.toml and Cargo.toml had
diverged and needed to be reunified.
Also the `alloy-signer-local` was unused and needed to be removed.

## Proposal

Both changes are done.

The elimination of `alloy-signer-local` is more important than it looks
like. Version 0.9.2 depends on `c-kzg`.
Unfortunately, version 22.0.1 of REVM depends on `c-kzg` as well and a
native library can be present only
one time in the linking process.

In other words, one block for the use of the next version of Revm has
been cleared. Two more obstacles remain:
* The resolution of the issue
https://github.com/bluealloy/revm/issues/2425
* The upgrade of Revm from version 22.0.1 to the next one with this bug
correction and the change of the PR
https://github.com/bluealloy/revm/pull/2464

When this is done, we could finally use Linera ApplicationId (the 20
bytes long version of course) in EVM contracts.

## Test Plan

The CI.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.