Implement the processing of the EVM gas usage. (#3715)
## Motivation

Running the EVM has its cost, known famously as gas/fuel. We need to
have adequate tracking of gas usage.

## Proposal

The following was done:
* For each operation done in storage, we track whether it maps from or
to 0 (meaning inexistent) and whether it changes the value. In a
separate set of tests, we identified when the corresponding gas costs.
* The identified gas costs are then subtracted from the gas_usage
returned by REVM. This is because the storage costs are being billed at
the end when the block is created. Double-counting the costs would be
inadequate.
* The `consume_fuel` and other are multiplied in order to take the
`VmRuntime` as arguments. The additional values are added to the
`Resources` and other entries.
* In a classical REVM run, the runs from separate contracts are grouped
together. However, we do not do that and run the different contracts
separately. Therefore, the other calls are set to zero. The grouping of
the different runs is then done in the `Resources` data types.
* The service tests are bounded in gas usage by the hardcoded value of
20_000_000.
* The metrics are corrected accordingly.

## Test Plan

The CI.

## Release Plan

Normal release of Linera.

## Links

None.