Address two issues with EVM (#3899)
## Motivation

We identified two problems with the EVM, which we correct now. Also,
some typos from the closed PR #3883 are put there.

## Proposal

The `gas_limit` is added to the BlobkEnv variable.

A more serious problem is the constructor argument for the contract
being run in `execute_message`. It was incorrect when running a
non-trivial constructor on another chain. The serialization was missing.

## Test Plan

A non-trivial constructor that does not do anything, just making a
check, was introduced to `evm_example_execute_message.sol`. It triggers
the previous problem.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

<!--
Optional section for related PRs, related issues, and other references.

If needed, please create issues to track future improvements and link
them here.
-->
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)