Add dual store support (#3878)
## Motivation

Our benchmarks are currently using `DualStore`, so it makes sense for us to add support to use it if we need it. We won't ideally need this, as the plan is to get `ScyllaDB` to a state that we reach the big numbers with it.

## Proposal

This PR adds the `linera-protocol` side support for `DualStore` networks, as well as support for doing local `linera net up --kubernetes` runs with `DualStore`

## Test Plan

Ran a local network with default params (scylladb) and `DualStore`, both work

## Release Plan

- Nothing to do / These changes follow the usual release cycle.