Remove 2 deprecated CLI options of `linera wallet init`. (#3910)
## Motivation

`--with-new-chain` and `--with-other-chains` are not necessary anymore;
`request-chain` and `follow-chain` should be used instead.

## Proposal

Remove them, and the `FaucetOption` enum.

Also don't print the latest hash anymore: This was meant to avoid
trusting the faucet or expired committees, but we already don't sync on
`wallet init` anymore, so it's in the wrong place. I created #3909 to
revisit this once we address the rest of the long-range attack issue.

## Test Plan

The tests were updated to use `request-chain` instead.

## Release Plan

- Before the next testnet, the Linera Manual needs to be updated.

## Links

- Closes https://github.com/linera-io/linera-protocol/issues/3399.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)