Provide EVM support for some linera features (#3847)
## Motivation

The Linera features need to be provided for EVM. Some of this is done
here.

## Proposal

The following is provided:
* validation_round
* chain_ownership
* chain_id
* application_creator_chain_id
* read_data_blob
* assert_data_blob_exists

The precompile is extended for that purpose and the handling of the
various cases is reorganized for clarity.

## Test Plan

A test for the features has been provided in the CI.

## Release Plan

The goal is to provide all this for TestNet 3.

## Links

None.