Use proper kube context and kube config in helmfile (#3917)
## Motivation

When deploying multiple networks in parallel, we need to make sure that helmfile is using the proper kube context and kube config, to prevent it from trying to deploy to an incorrect cluster.

## Proposal

Pass those in and use them when appropriate

## Test Plan

Deployed networks in parallel using local SSD, it works.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.