Remove unused code. (#3930)
## Motivation

Using
[cargo-workspace-unused-pub](https://crates.io/crates/cargo-workspace-unused-pub)
I found some unused code. Of course it is often legitimate to have
public code that is not used within the workspace, but I believe most of
the items it found are actually not needed.

## Proposal

Remove some unused code.

## Test Plan

CI—and a careful review so that I don't remove anything that is meant
for other users of the crates.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)