Use non admin chain for faucet if chain not specified (#3918)
## Motivation

Right now if you don't specify a chain, the faucet will just use the default chain of the wallet. However, for the faucet, this will often be using the admin wallet, which means the default chain will be the admin chain.

## Proposal

Instead of using the default chain, use a non admin chain instead in this case.

## Test Plan

Deployed a network with a faucet using this updated code, and it works as expected.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.