`linera-faucet`: return a whole `ChainDescription` (#3925)
## Motivation

Currently we work with the faucet by first claiming a chain, which
creates that chain's description in the network, then downloading the
chain description from the validators into the client storage. This adds
a roundtrip, but more importantly requires that we already have a client
before getting a chain from the faucet, making it hard to work with the
wallet before the client is initialized, which is necessary for e.g.
initializing signers.

## Proposal

Instead, return the whole chain description directly from the faucet,
and remove the `ClaimOutcome` type (and thus the `linera-faucet` crate,
which currently exists just to hold it).

## Test Plan

CI.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)