Remove chain ID from mutex before closing it. (#3953)
## Motivation

It looks like the faucet on the testnet closed a chain but then failed
to switch to a new one.

## Proposal

Remove the temporary chain ID from the mutex _before_ attempting to
close it and remove it from the wallet, so that in case any of those
steps fail, we will create a new temporary chain next time anyway.

Also, don't fail the request if we can't remove the chain from the
wallet.

## Test Plan

It's not clear yet what actually failed, but this change should be safe:
we could deploy it so that next time something goes wrong, the faucet
does not break entirely.

## Release Plan

- Deploy a new faucet on the testnet.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)