Enable access to the Ethereum Logs (#3937)
## Motivation

The Ethereum logs are an essential feature of the EVM, and we need to
support it. Since the events are available as blobs it is relatively
easy to support them.

Fixes #3758 

## Proposal

The following is done:
* The function `fn read_events_from_index` is added to the storage.
* Since the way the serialization is done influences the function,
adequate tests are added.
* The required traits are added to the `StreamId`,
`GenericApplicationId`, `StreamName`, and the needed tests are added so
that `FromStr` is the reverse of `Display`.
* The `IndexAndEvent` type is added, as this is how the events appear
from downloading them.
* GraphQL types are implemented via `SimpleObject` and `InputObject`
with `input_name` being used for consistency..
* The function `events_from_index` is added to the `NodeService`.
* The truncation length of 200 was not enough for debugging and was
extended to 500. It might be a good value overall, as 500 characters
will not flood the display.
* Several `.clone()` on storage that were not needed have been removed.
* The block height is added to the emitted events.


## Test Plan

The test for the feature has been added.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.