Revm alloy version bump (#3973)
## Motivation

We want to use the latest version of "alloy" and "revm" in Linera.
Upgrading to the most recent proved
far more difficult than expected.

## Proposal

The following changes are made:
* Both alloy and revm are upgraded simultaneously since there is some
issues with the package `c-kzg`.
* The `linera-ethereum` code is simplified by removing the `provider`
entry that was not needed.
* The REVM is changed to the most recent API which is quite different
from the one of 19.7.0


## Test Plan

The CI should do the job.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.