Add better OOM metrics to dashboards (#3975)
## Motivation

We've had OOMs in the shards, and this metric doesn't seem to have caught it.

## Proposal

Replace the graph with a graph using a metrics that _does_ seem to catch when a pod was `OOMKilled`. 

## Test Plan

I used this instead and saw the pods that were `OOMKilled` show there with the correct reason.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.