Account for published blobs in the resource at the end of transaction execution rather than upfront. (#3995)
## Motivation

We are charging for all published blobs before we execute any operation
in the block. This is wrong b/c we want the incoming grant to be able to
cover the fees. The grants are processed during operation processing
only so they are not taken into account at the beginning of the block
processing.

## Proposal

Move the published blobs' resources tracking to the end of processing a
transaction that published it. At this point any incoming grant is
already processed. In order to do so, we track blob IDs referenced
during transaction execution via `TransactionTracker.published_blobs`
and consume it in `BlockExecutionTracker::process_txn_outcome`.

Also, charge for publishing user-created blobs only.
`BlobType::Committee` is not accounted for.

## Test Plan

CI.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)