`linera-views`: centralize the `ViewError: From<Context::Error>` bound into `Context` (#4029)
## Motivation

We consistently require that the store errors be convertible to
`ViewError`, but we don't encode that bound in `Context`, so we
duplicate `ViewError: From<Context::Error>` bounds throughout the
codebase.

## Proposal

Move the bound into the definition of the `Context` trait instead.

While we're at it, use a boxed error for the store errors in `ViewError`
instead of converting to string, so that the backtrace and other error
metadata isn't lost.

## Test Plan

CI.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)