Change the API of `read_blob_state` to return options. (#4018)
## Motivation

Following PR https://github.com/linera-io/linera-protocol/pull/4014 we
look at the `read_blob_state` and change their API as options. The same
design considerations apply here.

## Proposal

The `read_blob_state(s)` were returning a ViewError that contained a
`NotFound`. This is inadequate and in any
case the `NotFound` is never processed.

In all use cases of those functions, we had some `BlobsNotFound` variant
in the relevant types, the error type was thus put there. It could
change the behavior of some function. However, this is better than
having some error buried in `ViewError`.

The `NotFound` cannot yet be changed because it is used by the
`read_certificate(s)`.

Other changes:
* After PR 4104, some of the metrics were placed before the storage
calls. This has been corrected.
* In some of the accounting of the calls, the plural calls were
sometimes counting several calls as one, and in other cases as the
number of entries in the call. This has been corrected.

## Test Plan

The CI.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.