`linera-views`: make the context an associated type (#4031)
## Motivation

Since I'm working with `linera-views` pretty heavily, I'd like to make
it easier to modify. One source of pain is that we parameterize the
`View` trait by its context type. This is semantically incorrect, as
each view can have only a single context type that is fully determined
by its fully-instantiated type (specifically, by the type of the context
contained within the view), and pragmatically speaking means we can't
write down the requirements on that context in the `View` trait itself,
and have to duplicate them at use sites, which makes modifying those
bounds painful.

## Proposal

Make the `Context` an associated type on the `View` trait, with all
necessary bounds (except some `Sync`s: see below).

## Test Plan

CI.

## Release Plan
- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)

## Future work

There remains a `V: View + Sync` bound on some of the implementers (e.g.
`MapView`) that forbids using those implementers on the Web, where views
are not `Sync` as the `IndexedDbStore` is not `Sync`. Removing this
bound is non-trivial because there is a `ReadKeyValueStore`
implementation for `storage-service` that is decidedly not `Sync`. I
will need to make that implementation `Sync`, remove the need for the
`Sync` bound, or do some complex conditional compilation, but this is at
least a step in the right direction.