Always add admin chain to listener. (#4044)
## Motivation

The chain listener currently panics if it has an empty list of chains.

Also, it doesn't even listen to the admin chain if no chain is in the
wallet.

## Proposal

Always listen to the admin chain.

## Test Plan

A unit test for this was added.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- Closes https://github.com/linera-io/linera-protocol/issues/4037.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)