Make signing with EmbeddedSigner work.
The root cause for problems here was serializing CryptoHash using
wasm_bindgen and serde. We can't serialize, because that adds new bytes
which are not present when verifying the signature. We are working with
the bytes - that CryptoHash wraps - directly. JsValue was a wrong type
to use as well. Now we use UInt8Array as that affects the logic inside
the ethers' signMessage method.

To summarize: CryptoHash had to be passed in as raw bytes it wraps.