Small changes to speed up and trim down CI costs. (#4032)
## Motivation

Speed up and cut down CI costs.

## Proposal

`docker-compose` was ran on every push which seems unnecessary. It
checks that our docker setup (that is expected to mimic validators'
usage) is working. Since we care that it works on `main` and possibly
release branches it doesn't have to run on every commit. **Now it runs
on merge queue (before merge to `main`) and on release branches**


`test-readme-scripts` tests backwards compatibility. This is important
but 95% of our PRs do not break them. We can detect it on `main` (and
release branches) and fix right away. **Now it runs on merge queue
(before merge to `main`) and on release branches**.

`lint-check-all-features` **runs on merge queue (before merge to `main`)
and on release branches**.

`default-features-and-witty-integration-test` now runs on
`ubuntu-latest-8-cores` : reason being this task already lasted 40+
minutes and is `cargo test` which should parallelize nicely.

`check-wit-files` runs on `ubuntu-latest` (rather than `*-16-cores`).

## Test Plan

CI testing CI.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.
## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)

---------

Signed-off-by: deuszx <95355183+deuszx@users.noreply.github.com>
Co-authored-by: Andreas Fackler <afck@users.noreply.github.com>