Fix merge queue required status. (#4061)
Due to limitations of GitHub's merge queues we can't set the different
"required workflows" for pull requests and merge queue. There are four
jobs we want to run on merge queue only but also require them to pass
before merging:
- docker compose
- Lint check all features
- Long Faucet chain test
- Run README tests

So now we add these to all commits but skip if the current CI run is NOT
merge queue. Now we can add them as required to pass for github
settings.