Fail earlier if re-proposal round is wrong. (#4090)
## Motivation

We never explicitly check that a re-proposal's round is actually greater
than the original proposal's. I couldn't find a way to actually exploit
it because the checks inside the chain manager seem to implicitly cover
all the cases, but it would be better to fail earlier, and it's a very
cheap check.

Thanks, @deuszx, for pointing this out!

## Proposal

Add this to `BlockProposal::check_invariants`.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)