TypeScript signer is a library/SDK  (#4105)
## Motivation

When building `linera-web` examples it sometimes happens that the
`package-lock.json` of `/signer` is incompatible/out of date. Both
signer and web client are used as SDK/libraries so should't have their
lock files committed.

`linera-web/build.bash` builds the client for the web but it didn't
prepare the TypeScript signers which required downstream client to
manually trigger that.

## Proposal

Signer and linera web client are a library/SDK and as such should not
have `package-lock.json` committed. This PR removes that file.

Add a line to `build.bash` that prepare signers (build javascript and
typescript files).

## Test Plan

Manually in `linera-web`.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links


- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)

---------

Signed-off-by: deuszx <95355183+deuszx@users.noreply.github.com>
Co-authored-by: James Kay <james.kay@linera.io>