Don't overwrite wallet entries on follow-chain. (#4135)
## Motivation

If the user accidentally calls `follow-chain` for an existing chain, it
overwrites the entry in the wallet.

## Proposal

Only write it if it doesn't exist yet.

## Test Plan

(It's not an important issue, so I don't think it's worth adding another
test for this edge case that slows down CI.)

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- Fixes https://github.com/linera-io/linera-protocol/issues/4132.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)