Only listen to faucet chain; not other chains in wallet. (#4136)
## Motivation

In the testnet we sometimes use the faucet with a wallet with lots of
chains. The chain listener then synchronizes them all, which is
wasteful.

## Proposal

Allow specifying a different set of chains for the chain listener, so
that it doesn't use all chains in the wallet.

In the faucet, start the chain listener with only the admin chain and
the faucet's main chain.

## Test Plan

We should now be able to start the testnet faucet with the original
wallet again, without it spending a lot of time downloading blocks.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)