Fix paths-ignore on pull_request when using merge_group (#4144)
## Motivation

It is a [known
problem](https://github.com/orgs/community/discussions/45899) that path
filtering doesn't work well when also using `merge_group`.
Basically if you have path filtering (like `paths-ignore`), in a
required workflow, if you do a change that gets filtered out by the path
filtering, that means that the required workflow will not run and the PR
will be blocked from being merged.
The documentation [states this
clearly](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/collaborating-on-repositories-with-code-quality-features/troubleshooting-required-status-checks#handling-skipped-but-required-checks).
One of the commenters in the issue suggested this workaround.
It's not the greatest thing as we need to add a check to every job in
the workflow now, but I'm not sure there's a better alternative.

## Proposal

Implement suggested workaround.

## Test Plan

PRs stacked on top of this one that were blocked are now no longer
blocked, and the runs were skipped as expected.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.