Force frame pointers in production docker image (#4140)
## Motivation

Quoting a presentation from Brendan Gregg:
> Once upon a tme, x86 had fewer registers, and the frame pointer
register was reused for general purpose to improve performance. This
breaks system stack walking.

Still, to this day, the default is for the frame pointer register to be
reused as a general purpose register, so frame information will be lost
as this register gets reused.

When using BPF-based debugging tools, since BPF runs in the kernel, it
doesn't have access to a DWARF unwinder, as the kernel doesn't have one
(and [a fully featured DWARF unwinder is unlikely to ever land in the
kernel](https://lkml.org/lkml/2012/2/10/129)). So in those cases we're
stuck with walking the stack using frame pointers.

## Proposal

Using the `-C force-frame-pointers=yes` Rust flag enforces that the
frame pointer register isn't reused for general purpose things, ensuring
that frame information is retained and we can properly see that
information in the stack traces.
[This should not have any significant impact on
performance.](https://fedoraproject.org/wiki/Changes/fno-omit-frame-pointer?#Potential_performance_impact)

## Test Plan

Have used this and the Docker image from the next PR while debugging our
current server memory issue.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.