Script to translate addresses in executable to file names and line numbers (#4141)
## Motivation

While using BPF's `memleak` to look into our server memory issue, I
could not for the life of me get it to automatically translate the PC
addresses into actual file names and file numbers.

## Proposal

Create a script that we can run with `memleak`'s output to translate
these addresses for us using `addr2line`. Once we figure out how to do
this in a more automated way (I'm assuming there must be a way), we can
remove this. But for now, it is useful for debugging.

## Test Plan

Have used this while debugging the memory issue in the server, the
stacks get translated correctly

## Release Plan

- Nothing to do / These changes follow the usual release cycle.