Add Dockerfile.debug for debugging purposes (#4142)
## Motivation

Production docker images are supposed to be as small as possible (more
work on shrinking ours is planned).
If there's a production issue for example, we don't want to bloat the
production Docker images by pre-installing a bunch of tools required for
debugging.
Also depending on how constrained the resources already are (depending
on the issue we're dealing with), installing different debugging tools
on the spot might not be feasible.

## Proposal

Introduce a debugging Docker image, which contains already most tools
one would generally use during debugging (other suggestions are
welcomed).
Instructions on how to use it are in the Dockerfile comments

## Test Plan

Have used this while debugging the server memory issue

## Release Plan

- Nothing to do / These changes follow the usual release cycle.