Restructure the calls to `read_certificate(s)`. (#4062)
## Motivation

The functions `read_certificate(s)` are introducing an entry
`ViewError::NotFound` in the `ViewError`, which is not great.
Here we eliminate this.

## Proposal

What we want to do is separate errors that are of storage access of pure
bugs from the errors that are more normal in operations.

There are 4 schemes of error in `deserialize_certificate`:
* Two `NotFound` patterns.
* One assertion that the value of the hash coincides.
* One check of consistency of the hashes.

Now they are grouped in two error patterns: `ReadCertificateError` and
`ReadCertificatesError`.
The resulting error type is something complicated of the form:
`Result<Result<Certificate, ReadCertificateError>, ViewError>`. The
corresponding reading is thus `??`.

This is the logic of what we do: First deal with the error of bad access
to the storage, and then with the error of the reading itself that was
not correct.

The corresponding `ReadCertificatesError` are added to the `WorkerError`
and `ChainClientError`.

This, of course, does not resolve the errors themselves. But exposing
them directly is better than having them burried in `ViewError`.

There are still patterns of errors `NotFound` left to eliminate.

## Test Plan

The CI.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.