Detect early if a validator sent the wrong certificates. (#4147)
## Motivation

We want to optimize how we download certificates and distribute work
among the validators in the future. To cancel requests to other
validators for the same certificate, we need to be sure that we got the
certificate we requested.

## Proposal

Compare the downloaded certificate's hash with the requested hash
immediately.

Also, remove a redundant argument from `download_certificates`: We
always use it with `validator_nodes()`.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)